igsc: lib: fix psc update
Fixed the wrong handling of meta-data in the PSC update.
The previous implementation wrongly assumed that for the PSC update
there is no meta-data part at all, while in reality the meta-data
structure should be included with all-zeros value.
The fix inserts the all-zeros meta-data structure for the PSC update.

Signed-off-by: Vitaly Lubart <vitaly.lubart@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>