igsc: lib: implement get subsystem ids library function
The igsc library used to get the subsystem vendor id (ssvid) and the
subsystem device id (ssdid) values from the PCIe config space of the
device on the host.

The firmware update library reads the ssvid/ssdid
from the firmware and not from the host.

This patch implements getring subsystem ids library function.

Signed-off-by: Vitaly Lubart <vitaly.lubart@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>