igsc: lib: initialize scalar variables
Initialize scalar variables in gsc_get_properties function and
received_len variable in various receive functions. The variables are
passed by address as parameters to functions that set their value, so
we need to inialize the variables to prevent a situation where the
functions does not set a value and so the variable stays uninitialized.

Signed-off-by: Vitaly Lubart <vitaly.lubart@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>