igsc: lib: remove check for version in get_hw_config
Some older projects like DG1 did not have get_hw_config heci
command implemented while the new ones like DG2 do implement it.
We were checking the version to determine whether this
command is supported by the firmware or not. The same check was
done to determine whether the image has hw_config data in it or not.
With more projects coming, those checks are not robust, we remove them
and instead check the value returned by the firmware which reports
NOT_SUPPORTED if the hw_config command is not implemented.
The logic for retrieval the hw_config from the image is changed
in the same way.

Signed-off-by: Vitaly Lubart <vitaly.lubart@intel.com>