igsc: lib: ignore return value of write on fwu_end command
There is a possible race condition between the firmware and the software
on writing the FWU_END command. When the command is written, the
firmware acknowledges the write interrupt and then immediately performs
a firmware reset. If the write interrupt ack wasn't processed quickly
enough by the host software driver - it may receive the interrupt of the
firmware reset together with the previous write interrupt ack.
In this case the igsc would assume that the write has failed.
Removing the check for the return value of the write of FWU_END command
to work-around this possible race condition.

Signed-off-by: Vitaly Lubart <vitaly.lubart@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>