igsc: lib: fix device enumeration in linux
Add check of return values of the udev functions and fail with
an appropriate error message in case of failure.
Free the enumeration when it's no longer needed, to prevent memory leak.
Call udev_device_get_sysname only once, and store the value,
no need to call it several times.

Signed-off-by: Vitaly Lubart <vitaly.lubart@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>